Scrapped quick and merge sort
I am unsure of how to implement these visually as they are recursive and each array that I "have" is not the full array unsorted or sorted, but a break-down of the full array. I have some ideas in mind but for now, these two sorts will be scrapped. Maybe they will come back, though!

Updated README for future endeavors.

Started on added more CSS to canvas (an overhaul of the visualizer coming?)